Extract placeholders to separate package
