+++ package.json
@@ -2,7 +2,7 @@
   "$schema": "https://www.raycast.com/schemas/extension.json",
   "name": "pins",
   "title": "Pins",
-  "version": "1.5.0",
+  "version": "1.6.0",
   "description": "Create pins for paths and URLs and display them in the menu bar",
   "keywords": [
     "pin",
@@ -393,6 +393,7 @@
     "@types/papaparse": "^5.3.8",
     "@types/xml-js": "^1.0.0",
     "papaparse": "^5.4.1",
+    "placeholders-toolkit": "^0.1.0",
     "xml-js": "^1.6.11",
     "yaml": "^2.3.2"
   },
